all reservation system done

+++ src/Entity/Taxe.php
@@ -19,7 +19,7 @@ class Taxe
     private $id;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $value;
 
@@ -55,12 +55,12 @@ class Taxe
         return $this->id;
     }
 
-    public function getValue(): ?int
+    public function getValue(): ?float
     {
         return $this->value;
     }
 
-    public function setValue(int $value): self
+    public function setValue(float $value): self
     {
         $this->value = $value;
